requirements.txt
anyascii==0.2.0
asgiref==3.3.4
astroid==2.7.2
beautifulsoup4==4.9.3
certifi==2021.5.30
charset-normalizer==2.0.4
cloudinary==1.26.0
colorama==0.4.4
dj-database-url==0.5.0
Django==3.2
django-cloudinary-storage==0.3.0
django-cors-headers==3.7.0
django-filter==2.4.0
django-heroku==0.3.1
django-modelcluster==5.1
django-taggit==1.5.1
django-treebeard==4.5.1
djangorestframework==3.12.4
draftjs-exporter==2.1.7
et-xmlfile==1.1.0
gunicorn==20.1.0
html5lib==1.1
idna==3.2
isort==5.9.3
l18n==2020.6.1
lazy-object-proxy==1.6.0
mccabe==0.6.1
openpyxl==3.0.7
Pillow==8.2.0
platformdirs==2.2.0
psycopg2==2.9.1
pylint==2.10.2
pylint-django==2.4.4
pylint-plugin-utils==0.6
pytz==2021.1
requests==2.26.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
tablib==3.0.0
telepath==0.2
toml==0.10.2
urllib3==1.26.6
wagtail==2.14.1
webencodings==0.5.1
whitenoise==5.3.0
Willow==1.4
wrapt==1.12.1
xlrd==2.0.1
XlsxWriter==1.4.5
xlwt==1.3.0
